Ajouts des résultats du scripts (CSV), première version finale du script

+++ main_script.py
@@ -4,6 +4,7 @@ import requests
 import pandas as pd
 from dotenv import load_dotenv
 from google.cloud import storage
+from datetime import datetime
 
 # Charger les variables d'environnement
 load_dotenv()
@@ -14,7 +15,7 @@ WM_API_KEY = os.getenv("WM_API_KEY")
 GCS_BUCKET_NAME = os.getenv("GCS_BUCKET_NAME")
 
 # Charger les villes depuis le fichier JSON
-with open("Projet_MSPR_5/ville_traitement.json", "r") as file:
+with open("C:/Users/camil/OneDrive - Ifag Paris/Cours/MSPR_EID_BLOC_5/Projet_MSPR_5/ville_traitement.json", "r") as file:
     VILLES = json.load(file)["villes"]
 
 # URLs des API
@@ -39,13 +40,14 @@ def fetch_weather_map(lat, lon, date):
         print(f"Erreur API WeatherMap pour {lat}, {lon}: {response.status_code}")
         return None
 
-def transform_air_quality(data):
+def transform_air_quality(data, city):
     if not data or data.get("status") != "ok":
         return None
     
     aq_data = data["data"]
     df = pd.DataFrame([{  
-        "city": aq_data["city"]["name"],
+        "city": city,
+        "Nom complet": aq_data["city"]["name"],
         "aqi": aq_data["aqi"],
         "pm25": aq_data["iaqi"].get("pm25", {}).get("v", None),
         "pm10": aq_data["iaqi"].get("pm10", {}).get("v", None),
@@ -78,7 +80,10 @@ def transform_weather_map(data, city, date):
     return df
 
 def save_csv(df, filename):
-    df.to_csv(filename, index=False)
+    if os.path.exists(filename):
+        df.to_csv(filename, mode='a', header=False, index=False)
+    else:
+        df.to_csv(filename, index=False)
     print(f"Données sauvegardées en local : {filename}")
 
 def save_to_gcs(df, filename):
@@ -90,6 +95,12 @@ def save_to_gcs(df, filename):
     print(f"Données sauvegardées dans GCS : {filename}")
 
 def main():
+    all_aq_data = []
+    all_weather_data = []
+    date_du_jour = datetime.today().strftime("%Y-%m-%d")
+
+    save_path = "C:/Users/camil/OneDrive - Ifag Paris/Cours/MSPR_EID_BLOC_5/Projet_MSPR_5/"
+    
     for ville in VILLES:
         city, lat, lon = ville["nom"], ville["lat"], ville["lon"]
         
@@ -98,18 +109,29 @@ def main():
         weather_data = fetch_weather_map(lat, lon, "2025-02-12")  # Exemple avec une date fixe
         
         # Transformation des données
-        aq_df = transform_air_quality(aq_data)
-        weather_df = transform_weather_map(weather_data, city, "2025-02-12")
+        print(f"Transformation des données de Air Quality pour : {city}")
+        aq_df = transform_air_quality(aq_data, city)
+        print(f"Transformation des données de WeatherMap pour : {city}")
+        weather_df = transform_weather_map(weather_data, city, date_du_jour)
         
-        # Stockage sur GCS
+        # Ajout aux listes
         if aq_df is not None:
-            filename = f"air_quality_{city}.csv"
-            save_csv(aq_df, filename)
-            # save_to_gcs(aq_df, filename)
+            print(f"Ajout aux listes de air quality pour {city}")
+            all_aq_data.append(aq_df)
         if weather_df is not None:
-            filename = f"weather_{city}.csv"
-            save_csv(weather_df, filename)
-            # save_to_gcs(weather_df, filename)
+            print(f"Ajout aux listes de weathermap pour {city}")
+            all_weather_data.append(weather_df)
+    
+    # Sauvegarde finale
+    if all_aq_data:
+        final_aq_df = pd.concat(all_aq_data, ignore_index=True)
+        save_csv(final_aq_df, f"{save_path}Qualite_Air.csv")
+        # save_to_gcs(final_aq_df, "Qualite_Air.csv")
+    
+    if all_weather_data:
+        final_weather_df = pd.concat(all_weather_data, ignore_index=True)
+        save_csv(final_weather_df, f"{save_path}Ville_Meteo.csv")
+        # save_to_gcs(final_weather_df, "Ville_Meteo.csv")
 
 if __name__ == "__main__":
     main()
